requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-green==3.6.6
aliyun-python-sdk-kms==2.16.3
Django==5.0.7
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
oss2==2.18.6
mysqlclient==2.1.1
PyJWT==2.3.0
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
redis==5.0.7
requests==2.25.1
torch==2.4.0+cpu
